+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.15.0-dev
 - added RSS extensions Comment API and blogChannel
 - BaseElement::registerElements() can now set elements' types, normalize enum values and mark elements as required
 - only first letter is now changed to lower case in BaseExtension::getName()
+- made prefix for elements' constants for BaseElement::registerElements() customizable
 
 Version 0.14.0
 - added RSS extension CreativeCommons