+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 0.15.0-dev
+Version 0.15.0
 - added RSS extensions Comment API and blogChannel
 - BaseElement::registerElements() can now set elements' types, normalize enum values and mark elements as required
 - only first letter is now changed to lower case in BaseExtension::getName()